+++ velocyto/__init__.py
@@ -8,7 +8,6 @@ from .indexes import FeatureIndex, TransciptsIndex
 from .molitem import Molitem
 from .logic import *
 from .counter import ExInCounter
-from .neg_ctrl_counter import FlippingCounter, PolyTCounter
 from .metadata import MetadataCollection, Metadata
 from .neighbors import BalancedKNN, convolve_by_sparse_weights
 from .estimation import fit_slope, _fit1_slope, clusters_stats