+++ doc/changelog.rst
@@ -6,6 +6,7 @@ Changelog
 
 * :release:`0.12.0 <2017-12-17>`
 * :feature:`-` Add possibility to export pickle containing information of every molecule
+* :feature:`-` Remove the subcommand ``multi10x``
 * :bug:`-` Incorrect 0-based indexing for splicing junction corrected (was not causing problems because buffered by MIN_FLANK) 
 * :bug:`-` Many small bug fixes
 * :feature:`-` Large parts of the documentation rewritten to match the changes in API